Revision Session, Test functionning over Run button but not in terminal

+++ collaborator_app/controller.py
@@ -1,11 +1,11 @@
+import logging
 from client_app.view import ClientView
-from services.utils import CollaboratorService, AuthenticateService, CollaboratorModificationService
-from collaborator_app.model import Collaborator, PH
-from client_app.model import Client
-from event_app.model import Event
+from services.utils import CollaboratorService, AuthenticateService, CollaboratorModificationService, Session
+from collaborator_app.model import Collaborator
 from collaborator_app.view import MenuView, display_collaborator_table
 from client_app.controller import MainClientController
 from contract_app.controller import MainContractController
+from event_app.controller import MainEventController
 
 
 class MainController:
@@ -14,24 +14,17 @@ class MainController:
 
     @staticmethod
     def run():
-        token = AuthenticateService.load_token()
-        if token:
-            user_id = AuthenticateService.decode_token(token)
-            collaborator = CollaboratorService.get_collaborator_by_id(user_id)
-            if collaborator:
-                MainController.display_role_menu(collaborator.role)
-                return
+        if Session.load_session():
+            MainController.display_role_menu(Session.user.role)
+            return
 
         while True:
             MenuView.display_menu()
             choice = MenuView.get_choice()
             if choice == MainController.LOGIN:
                 username, password = MenuView.display_login_menu()
-                collaborator = CollaboratorService.authenticate_collaborator(username, password)
-                if collaborator:
-                    token = AuthenticateService.generate_token(collaborator.id)
-                    AuthenticateService.store_token(token)
-                    MainController.display_role_menu(collaborator.role)
+                if Session.authenticate(username, password):
+                    MainController.display_role_menu(Session.user.role)
                     break
                 else:
                     MenuView.display_fail_auth()
@@ -41,26 +34,17 @@ class MainController:
 
     @staticmethod
     def display_existing_collaborators():
-        """Fonction qui récupère les champs souhaité pour l'affichage des collaborateurs dans les menus en indiquant en params de select les champs voulus"""
-        collaborators = Collaborator.select(Collaborator.id,
-                                            Collaborator.username,
-                                            Collaborator.first_name,
-                                            Collaborator.last_name,
-                                            Collaborator.role)
+        collaborators = Collaborator.select()
         MenuView.display_collaborators(collaborators)
 
     @staticmethod
     def display_role_menu(role):
-        if role == Collaborator.GESTION:
-            MainController.gestion_control_menu()
-        elif role == Collaborator.COMMERCIAL:
+        if role == Collaborator.COMMERCIAL:
             MainController.commercial_control_menu()
+        elif role == Collaborator.GESTION:
+            MainController.gestion_control_menu()
         elif role == Collaborator.SUPPORT:
-            MenuView.display_support_menu()
-            choice = MenuView.get_choice()
-            if choice.upper() == "Q":
-                AuthenticateService.delete_token()
-                MainController.run()
+            MainController.support_control_menu()
 
     @classmethod
     def gestion_control_menu(cls):
@@ -74,8 +58,14 @@ class MainController:
                 MainClientController.display_existing_client()
             elif choice == "3":
                 MainContractController.display_existing_contract()
+            elif choice == "4":
+                MainEventController.display_existing_event()
             elif choice == "5":
                 MainContractController.create_contract()
+            elif choice == "6":
+                MainContractController.modify_contract()
+            elif choice == "7":
+                MainEventController.modify_event()
             elif choice == "8":
                 MainController.display_existing_collaborators()
                 cls.modify_collaborator()
@@ -95,7 +85,51 @@ class MainController:
             choice = ClientView.get_choice()
             if choice == "1":
                 MainClientController.create_client()
+            elif choice == "2":
+                MainClientController.display_existing_client()
+            elif choice == "3":
+                MainContractController.display_existing_contract()
+            elif choice == "4":
+                MainEventController.display_existing_event()
+            elif choice == "5":
+                MainClientController.modify_client()
+            elif choice == "6":
+                MainContractController.modify_contract()
+            elif choice == "7":
+                MainContractController.display_contracts()
+            elif choice == "8":
+                MainEventController.create_event()
+            elif choice == "9":
+                MainController.display_existing_collaborators()
             elif choice.upper() == "Q":
+                AuthenticateService.delete_token()
+                MainController.run()
+
+    @classmethod
+    def support_control_menu(cls):
+        while True:
+            MenuView.display_support_menu()
+            choice = ClientView.get_choice()
+            if choice == "1":
+                logging.info("display_existing_client called")
+                MainClientController.display_existing_client()
+            elif choice == "2":
+                logging.info("display_existing_contract called")
+                MainContractController.display_existing_contract()
+            elif choice == "3":
+                logging.info("display_existing_event called")
+                MainEventController.display_existing_event()
+            elif choice == "4":
+                logging.info("display_existing_collaborators called")
+                MainController.display_existing_collaborators()
+            elif choice == "5":
+                logging.info("list_events called")
+                MainEventController.list_events()
+            elif choice == "6":
+                logging.info("modify_event called")
+                MainEventController.modify_event()
+            elif choice.upper() == "Q":
+                logging.info("Se déconnecter called")
                 AuthenticateService.delete_token()
                 MainController.run()
 
@@ -113,26 +147,32 @@ class MainController:
 
     @classmethod
     def modify_collaborator(cls):
+        if not Session.is_gestion():
+            MenuView.display_insufficient_permissions()
+            return
+
         collaborator_id = MenuView.modify_collaborator_by_id()
         data = MenuView.get_new_collaborator_info()
         CollaboratorModificationService.validate_collaborator_modification_data(collaborator_id)
-
         collaborator = CollaboratorService.get_collaborator_by_id(collaborator_id)
         collaborator.set_password(data['password'])
         collaborator.username = data["username"]
         collaborator.first_name = data["first_name"]
         collaborator.last_name = data["last_name"]
         collaborator.role = data["role"]
-        Collaborator.save(collaborator)
-        token = AuthenticateService.load_token()
-        if token:
-            user_id = AuthenticateService.decode_token(token)
-            if user_id == collaborator_id:
-                AuthenticateService.delete_token()
-                MainController.run()
+
+        collaborator.save()
+        current_user = Session.get_current_user()
+        if current_user.id == collaborator_id:
+            Session.logout()
+            MainController.run()
 
     @classmethod
     def delete_collaborator(cls):
+        if Session.user.role != Collaborator.GESTION:
+            MenuView.display_insufficient_permissions()
+            return
+
         collaborator_id = MenuView.delete_collaborator_by_id()
         try:
             collaborator = Collaborator.get_by_id(collaborator_id)
